Fix compiler warnings.
Add demo.menu and all xbm and xpm files in test directory to install
list.


git-svn-id: file:///fltk/svn/fltk/branches/branch-1.1@4292 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fluid/Fl_Function_Type.cxx
@@ -551,13 +551,13 @@ void Fl_Decl_Type::write_code1() {
   while (e>c && e[-1]==';') e--;
   if (class_name(1)) {
     write_public(public_);
-    write_h("  %.*s;\n", e-c, c);
+    write_h("  %.*s;\n", (int)(e-c), c);
   } else {
     if (public_) {
-      write_h("extern %.*s;\n", e-c, c);
-      write_c("%.*s;\n", e-c, c);
+      write_h("extern %.*s;\n", (int)(e-c), c);
+      write_c("%.*s;\n", (int)(e-c), c);
     } else {
-      write_c("static %.*s;\n", e-c, c);
+      write_c("static %.*s;\n", (int)(e-c), c);
     }
   }
 }
